+++ app/src/main/java/org/simple/clinic/setup/SetupActivityEffect.kt
@@ -1,5 +1,7 @@
 package org.simple.clinic.setup
 
+import org.simple.clinic.appconfig.Country
+import org.simple.clinic.appconfig.Deployment
 import org.simple.clinic.setup.runcheck.Disallowed
 
 sealed class SetupActivityEffect
@@ -21,3 +23,10 @@ object FetchDatabaseMaintenanceLastRunAtTime : SetupActivityEffect()
 data class ShowNotAllowedToRunMessage(val reason: Disallowed.Reason) : SetupActivityEffect()
 
 object CheckIfAppCanRun : SetupActivityEffect()
+
+data class SaveCountryAndDeployment(
+    val country: Country,
+    val deployment: Deployment
+): SetupActivityEffect()
+
+object DeleteStoredCountryV1 : SetupActivityEffect()