+++ app/src/test/java/org/simple/clinic/appconfig/AppConfigRepositoryTest.kt
@@ -29,6 +29,7 @@ class AppConfigRepositoryTest {
   private val selectedDeployment = mock<Preference<Optional<Deployment>>>()
   private val selectedStatePreference = mock<Preference<Optional<String>>>()
   private val statesFetcher = mock<StatesFetcher>()
+  private val countryV1Preference = mock<Preference<Optional<String>>>()
 
   private val repository = AppConfigRepository(
       manifestFetchApi,
@@ -36,6 +37,7 @@ class AppConfigRepositoryTest {
       selectedDeployment,
       selectedStatePreference,
       statesFetcher,
+      countryV1Preference
   )
 
   @Test